Added: settings. fixed: #53, #53, several others
Version bumb to 0.25

Build in a function that will make sure everymodmail in /unread/ is
marked read when visiting a modmail related page. This should solve all
issues people have with this.

Using false as default value for TBUtils.setting would break the
function since it checked for the existence of a defaultvalue by
if(defaultvalue), changed this to if(defaultvalue !== undefined).

Added settings for links to /unread/ since some people do want to go
there and others don't. Might as well make it a setting then.

Moved icons to TBUtils.

Consolidated message notification will not turn up for zero messages
anymore.