Less spammy modmail with individual messages on
Since modmail is now checked based on the local system it will show messages already seen and read on other computers. When individual messages are enabled this can become a huge amount of popups. This change will automatically switch to consolidated messages if the amount of new modmails is > 6. I decided on 6 since there are three messages show at the time two rows of messages seems like a reasonable amount. In the future this likely has to become a setting.