obscure bug for removal reasons nobody will ever notice anyway.
The long version is that it wouldn't display "none" and the other log reason stuff if there was no {reason}  defined. It would still allow the log to be posted etc. This did happen to me in a test subreddit where I had once a log sub defined but completely forgotten about it. But since there was a log sub defined it would still allow none to be selected as the default. Which made me think the radioboxes were broken.
This allows for the obscure use case of people using this to log to a subreddit without ever putting in a reason.