Add missing syntax specifier to user history table
To use structured queries, you need to either write them in the default syntax (lucene), or specify that it's written in the alternate syntax, cloudsearch.  Right now, the tables in user history comments (at least in /r/spam reports) use cloudsearch but don't specify the syntax as such.  This technically works but might cause issues.

This only matters on those specific queries which have queries that use cloudsearch.  The other ones (on lines 460 and 660), do not use cloudsearch and thus do not need to be changed.