Beter centering and spacing in modbar
The modbar's contents are more consistently centered and spaced out. No
longer uses &nbsp; for spacing in the HTML.

To do the things, the HTML layout of the modbar changed to have 2 content
spans - one for the left, one for the right. Most things have been moved
inside these, and hopefully it shuoldn't break much since most things
interacting with the modbar used `.before()` rather than `.append()` and
the IDs of the actual things haven't changed.