Fixes #802
The lookup for stickied submissions wasn't specific enough and maybe when the entire module is about functionality in subs people mod it should have a check for that before everything is called. You know, because of performance and all that. 

Also, we have code guidelines for whoever made this atrocity of an init call. I don't care that it takes extra lines, we put {} around our functions for readability sake and because this isn't python.