Optimize watchPushState on old Reddit
The previous solution required animation frames to be scheduled as
long as the page was visible, preventing the CPU from idling.

This change adds a passive listener in the background which send a
message to the content script when the tab has a new URL. The
extension API is used as there's no native Web API which triggers an
event when the URL is modified by history#pushState.