Update modbar.js because new.reddit is dead
Replaced redirects to the now-defunct new.reddit to sh.reddit in the modbar "open in new reddit" button.

As of December 11th 2024, Reddit deleted new.reddit outright so the switching functionality was broken with this button. sh.reddit is the next best thing for this purpose so mods can still access the built-in (non-toolbox) removal reasons and other things inaccessible on old reddit.

For the record, new.reddit just goes straight to www.reddit now (whichever is configured for your account, be it "old" or "sh").